move app into test common file
