+++ server/tests/common.integration.js
@@ -4,3 +4,4 @@ import config from '../../config/config';
 const version = p.version.split('.').shift();
 export const baseURL = (version > 0 ? `/api/v${version}` : '/api');
 export const auth = config.testToken;
+export { default as app } from '../../index';
